backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
PyJWT
bcrypt
python-multipart
